pyproject.toml
[project]
name = "tracery-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.4.0",
    "guidance>=0.2.0",
    "llama-cpp-python>=0.3.7",
    "ninja>=1.11.1.3",
    "outlines",
    "sentencepiece>=0.2.0",
    "setuptools>=75.8.2",
    "torch==2.6.0+cu126",
    "transformers>=4.49.0",
    "xgrammar>=0.1.15",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch" },
]
torchvision = [
  { index = "pytorch" },
]
